Pegs, Masscroft, RAMON
Added new quest object for Sky Spy: Trangulation peg. Removed on activation of the gold thingys.
Connected the two scripts that control Masscroft battle so the script that controls his death (on trigger box) calls the command UnregisterForUpdate() on the script on Masscroft himself
Moved the StopTranslation() command for RAMON's hands and head to the top of the conditional statement for his death, so it's the very first thing that happens and no errors will come up. May not work because it's in two separate scripts and everything happens in milliseconds, no pun intended.