Stonehole Quest Marker
Added stage to MQ01 that reads "Travel to Stonehole Mine in The Pale" and points to Stonehole's door.
Removed ability to immidiatly fast travel to stonehole upon getting the objective to do so
The objective "reach the bottom of stonehole mine" no longer has a quest marker.
Made Numinar's door set to locked, requires key instead of novice