Eye of Sybandis random jumpscare
Places a few triggerboxes around the mod. When you enter the box after picking up the Idiot's Lantern, there is a 1/400 chance the Eye of Sybandis will appear and play a loud noise before vanishing.
Once the triggerbox is entered it is disabled, even if the Eye wasn't spawned, so actually seeing it will be INSANELY rare.
Feel free to make a tester MCM option to guarantee it spawns if you feel like it.