Stonehole Navmesh fix (sorry mayor. Save for 5.2?)
Had to balance the main shaft and clean it up to fix some cases where followers were setting stuck on platforms, meaning the Ck did the stupid deletion thing again. It's nothing serious so if you want to just hold onto this and make it part of the overall Stonehole navmesh cleanup I want to do for 5.2, that's totally fine.