Whitehorn Pass Tweaks
Move Secarius to the right side of the gate. Also slightly enlarge the setstage trigger so you can't miss it.
Note: I'm not sure what that xmarker linked to Secarius does, so I don't know what side of the gate it should be on. I left it on the other.