[Merge + Addition] Whitehorn Pass Tweaks
Move Secarius to the right side of the gate. Also slightly enlarge the setstage trigger so you can't miss it.
Note: I'm not sure what that xmarker linked to Secarius does, so I don't know what side of the gate it should be on. I left it on the other

Mayor addition: the xmarker is used by his default package, which is to hold position at a linked reference. So I moved it to the other side of the gate, and moved it and secarius a little to the side so the middle of the gate isn't blocking his face.

Co-Authored-By: DarthVitrial <DarthVitrial@users.noreply.github.com>